requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
langchain==0.1.16
flask==3.0.2
pypdf==4.0.2
python-dotenv==1.0.1
qdrant-client==1.7.3
langchain-community==0.0.29
torch==2.2.2
transformers==4.38.2
accelerate==0.27.2

-e .